pyproject.toml
[project]
name = "vision-desktop-automation"
version = "0.1.0"
description = "Vision-based desktop automation with dynamic icon grounding for Windows"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyautogui>=0.9.54",
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "pillow>=9.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

